feat: add new extractor plugins and update documentation
- Introduce `CopyContentsPlugin` for copying file contents into documentation
- Add `CallbackExtractorPlugin` for custom extraction logic using callbacks
- Implement `RegexExtractorPlugin` for regex-based content extraction
- Update `TagExtractorPlugin` to enhance tag-based extraction capabilities
- Revise documentation to include new plugins and usage examples
- Refactor imports to align with new plugin structure
