+++ src/simple-docs-scraper/plugins/CallbackExtractorPlugin.ts
