+++ src/simple-docs-scraper/plugins/CopyContentsPlugin.ts
