+++ src/simple-docs-scraper/plugins/DocumentContentExtractor.ts
