+++ src/simple-docs-scraper/plugins/RegexExtractorPlugin.ts
