+++ src/simple-docs-scraper/transformers/ContentInjection.ts
@@ -1,6 +1,6 @@
 import fs from "fs";
 import path from "path";
-import { ExtractedContent } from "../extractors/DocumentContentExtractor.js";
+import { ExtractedContent } from "../plugins/DocumentContentExtractor.js";
 import { Target } from "../services/SimpleDocExtractor.js";
 import { ExtractorPlugin } from "../types/extractor.t.js";
 import TemplateContentExtractionContentMerger from "./TemplateContentExtractionContentMerger.js";
